don't change UART settings on Stretch and newer

+++ upd_script/update_all.sh
@@ -328,17 +328,23 @@ sudo sed -i "/dtparam=spi=on/d" /boot/config.txt
 sudo echo "dtparam=spi=on" >> /boot/config.txt
 sudo echo "dtparam=i2c_arm=on" >> /boot/config.txt
 
-# This is really imprtant for the BrickPi!
-sudo sed -i "/init_uart_clock=32000000/d" /boot/config.txt
-sudo echo "init_uart_clock=32000000" >> /boot/config.txt
 
-# Disable serial over UART
-sudo sed -i 's/console=ttyAMA0,115200//' /boot/cmdline.txt  #disable serial login on older images
-sudo sed -i 's/console=serial0,115200//' /boot/cmdline.txt  #disable serial login on the Pi3
-sudo sed -i 's/console=tty1//' /boot/cmdline.txt            #console=tty1 can also be there in the cmdline.txt file so remove that
-sudo sed -i 's/kgbdoc=ttyAMA0,115200//' /boot/cmdline.txt
-sudo systemctl stop serial-getty@ttyAMA0.service
-sudo systemctl disable serial-getty@ttyAMA0.service
+# Only reset UART on Jessie as we still support the BrickPi+
+if [ $VERSION -eq '8' ]
+then
+    # This is really imprtant for the BrickPi!
+    sudo sed -i "/init_uart_clock=32000000/d" /boot/config.txt
+    sudo echo "init_uart_clock=32000000" >> /boot/config.txt
+
+
+    # Disable serial over UART
+    sudo sed -i 's/console=ttyAMA0,115200//' /boot/cmdline.txt  #disable serial login on older images
+    sudo sed -i 's/console=serial0,115200//' /boot/cmdline.txt  #disable serial login on the Pi3
+    sudo sed -i 's/console=tty1//' /boot/cmdline.txt            #console=tty1 can also be there in the cmdline.txt file so remove that
+    sudo sed -i 's/kgbdoc=ttyAMA0,115200//' /boot/cmdline.txt
+    sudo systemctl stop serial-getty@ttyAMA0.service
+    sudo systemctl disable serial-getty@ttyAMA0.service
+fi
 
 feedback "--> End Kernel Updates."
